# contracts/openapi.bundled.yaml
openapi: 3.0.3
info:
  title: Backend API
  description: RESTful API with authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: 'http://localhost:8080/api/v1'
    description: Development server
  - url: 'https://api.example.com/v1'
    description: Production server
...
paths:
...
  /patient-checkups:
    get:
      operationId: listPatientCheckups
      summary: Get patient checkup history
      description: Retrieve a paginated list of patient checkups
      tags:
        - patient_checkups
      security:
        - BearerAuth:
            - admin
            - doctor
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/PatientCheckupSearchParam'
        - $ref: '#/components/parameters/PatientCheckupPatientIdParam'
        - $ref: '#/components/parameters/PatientCheckupStatusParam'
        - $ref: '#/components/parameters/PatientCheckupVisitDateParam'
        - $ref: '#/components/parameters/PatientCheckupDateFromParam'
        - $ref: '#/components/parameters/PatientCheckupDateToParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientCheckup'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createPatientCheckup
      summary: Create patient visit schedule
      description: Admin creates patient visit schedule with initial symptoms
      tags:
        - patient_checkups
      security:
        - BearerAuth:
            - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientCheckupRequest'
      responses:
        '201':
          description: Patient checkup created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PatientCheckup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  '/patient-checkups/{id}':
    get:
      operationId: getPatientCheckup
      summary: Get patient checkup by ID
      description: Retrieve a specific patient checkup by UUID
      tags:
        - patient_checkups
      security:
        - BearerAuth:
            - admin
            - doctor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PatientCheckup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updatePatientCheckup
      summary: Update patient checkup
      description: 'Admin or doctor edits visit schedule, symptoms, and clinical result'
      tags:
        - patient_checkups
      security:
        - BearerAuth:
            - admin
            - doctor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientCheckupRequest'
      responses:
        '200':
          description: Patient checkup updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PatientCheckup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deletePatientCheckup
      summary: Delete patient checkup
      description: Delete a patient checkup record
      tags:
        - patient_checkups
      security:
        - BearerAuth:
            - admin
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Patient checkup deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
      example: 123e4567-e89b-12d3-a456-426614174000
...
    PatientCheckupSearchParam:
      name: search
      in: query
      schema:
        type: string
      description: |
        Search patient checkups by chief complaint, diagnosis, notes, doctor name, or symptoms (partial match, case-insensitive)
    PatientCheckupPatientIdParam:
      name: patient_id
      in: query
      schema:
        type: string
        format: uuid
      description: Filter checkup history by patient UUID
    PatientCheckupStatusParam:
      name: status
      in: query
      schema:
        type: string
        enum:
          - scheduled
          - completed
          - cancelled
      description: Filter checkups by workflow status
    PatientCheckupVisitDateParam:
      name: visit_date
      in: query
      schema:
        type: string
        format: date
      description: Filter checkups by visit date (exact day)
    PatientCheckupDateFromParam:
      name: visit_date_from
      in: query
      schema:
        type: string
        format: date
      description: Filter checkups with visit date from (>=)
    PatientCheckupDateToParam:
      name: visit_date_to
      in: query
      schema:
        type: string
        format: date
      description: Filter checkups with visit date to (<=)
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Invalid input
            code: BAD_REQUEST
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Authentication required
            code: UNAUTHORIZED
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Resource not found
            code: NOT_FOUND
...
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Invalid input
        code:
          type: string
          example: INVALID_INPUT
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Meta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 10
...
    PatientCheckup:
      type: object
      required:
        - id
        - patient_id
        - visit_date
        - status
        - chief_complaint
        - symptoms
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: 223e4567-e89b-12d3-a456-426614174111
          description: Checkup UUID
        patient_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
          description: Patient UUID
        visit_date:
          type: string
          format: date-time
          example: '2026-03-02T10:15:00Z'
          description: Planned visit date and time
        status:
          type: string
          enum:
            - scheduled
            - completed
            - cancelled
          example: scheduled
          description: Visit workflow status
        chief_complaint:
          type: string
          example: Demam sejak 2 hari
          description: Main complaint reported when booking
        symptoms:
          type: array
          items:
            type: string
          example:
            - demam
            - batuk
            - pusing
          description: Symptoms recorded and editable by admin/doctor
        diagnosis:
          type: string
          nullable: true
          example: Infeksi saluran pernapasan atas
          description: Clinical diagnosis
        temperature_c:
          type: number
          format: float
          minimum: 30
          maximum: 45
          nullable: true
          example: 38.4
          description: Body temperature in celsius
        blood_pressure:
          type: string
          nullable: true
          example: 120/80
          description: Blood pressure reading
        heart_rate:
          type: integer
          minimum: 0
          nullable: true
          example: 86
          description: Heart rate (bpm)
        respiratory_rate:
          type: integer
          minimum: 0
          nullable: true
          example: 18
          description: Respiratory rate (per minute)
        oxygen_saturation:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          example: 98
          description: SpO2 percentage
        height_cm:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 170
          description: Height in cm
        weight_kg:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 65.5
          description: Weight in kg
        medicines:
          type: array
          items:
            $ref: '#/components/schemas/PatientCheckupMedicine'
          description: Prescribed medicines during this checkup
        treatment_plan:
          type: string
          nullable: true
          example: 'Istirahat cukup, minum obat teratur, kontrol 3 hari'
          description: Treatment and follow-up plan
        notes:
          type: string
          nullable: true
          example: Pasien memiliki riwayat alergi penicillin
          description: Additional notes
        doctor_name:
          type: string
          nullable: true
          example: dr. Andi Pratama
          description: Doctor handling the checkup
        follow_up_date:
          type: string
          format: date
          nullable: true
          example: '2026-03-05'
          description: Planned follow-up date
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    PatientCheckupMedicine:
      type: object
      required:
        - medicine_id
        - medicine_name
        - quantity
        - dosage
        - frequency
        - duration_days
      properties:
        medicine_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
          description: Medicine UUID
        medicine_name:
          type: string
          example: Paracetamol
          description: Snapshot medicine name when prescribed
        quantity:
          type: integer
          minimum: 1
          example: 10
          description: Dispensed quantity from inventory
        dosage:
          type: string
          example: 500 mg
          description: Single dose amount
        frequency:
          type: string
          example: 3x sehari
          description: Frequency instruction
        duration_days:
          type: integer
          minimum: 1
          example: 5
          description: Duration in days
        notes:
          type: string
          nullable: true
          example: Setelah makan
          description: Additional usage notes
    CreatePatientCheckupRequest:
      type: object
      required:
        - patient_id
        - visit_date
        - chief_complaint
        - symptoms
      properties:
        patient_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        visit_date:
          type: string
          format: date-time
          example: '2026-03-02T10:15:00Z'
        chief_complaint:
          type: string
          example: Demam sejak 2 hari
        symptoms:
          type: array
          items:
            type: string
          example:
            - demam
            - batuk
            - pusing
        notes:
          type: string
          nullable: true
          example: Pasien minta jadwal pagi
        status:
          type: string
          enum:
            - scheduled
            - completed
            - cancelled
          example: scheduled
        diagnosis:
          type: string
          nullable: true
          example: Infeksi saluran pernapasan atas
        temperature_c:
          type: number
          format: float
          minimum: 30
          maximum: 45
          nullable: true
          example: 38.4
        blood_pressure:
          type: string
          nullable: true
          example: 120/80
        heart_rate:
          type: integer
          minimum: 0
          nullable: true
          example: 86
        respiratory_rate:
          type: integer
          minimum: 0
          nullable: true
          example: 18
        oxygen_saturation:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          example: 98
        height_cm:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 170
        weight_kg:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 65.5
        medicines:
          type: array
          items:
            $ref: '#/components/schemas/PatientCheckupMedicine'
        treatment_plan:
          type: string
          nullable: true
          example: 'Istirahat cukup, minum obat teratur, kontrol 3 hari'
        doctor_name:
          type: string
          nullable: true
          example: dr. Andi Pratama
        follow_up_date:
          type: string
          format: date
          nullable: true
          example: '2026-03-05'
        patient_allergies:
          type: string
          nullable: true
          example: 'Penicillin, seafood'
          description: Optional update for patient allergies from doctor's assessment
        patient_blood_type:
          type: string
          nullable: true
          enum:
            - A+
            - A-
            - B+
            - B-
            - AB+
            - AB-
            - O+
            - O-
          example: O+
          description: Optional update for patient blood type from doctor's assessment
    UpdatePatientCheckupRequest:
      type: object
      required:
        - visit_date
        - chief_complaint
        - symptoms
        - status
      properties:
        visit_date:
          type: string
          format: date-time
          example: '2026-03-02T10:15:00Z'
        chief_complaint:
          type: string
          example: Demam sejak 2 hari
        symptoms:
          type: array
          items:
            type: string
          example:
            - demam
            - batuk
            - pusing
        status:
          type: string
          enum:
            - scheduled
            - completed
            - cancelled
          example: scheduled
        diagnosis:
          type: string
          nullable: true
          example: Infeksi saluran pernapasan atas
        temperature_c:
          type: number
          format: float
          minimum: 30
          maximum: 45
          nullable: true
          example: 38.4
        blood_pressure:
          type: string
          nullable: true
          example: 120/80
        heart_rate:
          type: integer
          minimum: 0
          nullable: true
          example: 86
        respiratory_rate:
          type: integer
          minimum: 0
          nullable: true
          example: 18
        oxygen_saturation:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          example: 98
        height_cm:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 170
        weight_kg:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 65.5
        medicines:
          type: array
          items:
            $ref: '#/components/schemas/PatientCheckupMedicine'
        treatment_plan:
          type: string
          nullable: true
          example: 'Istirahat cukup, minum obat teratur, kontrol 3 hari'
        doctor_name:
          type: string
          nullable: true
          example: dr. Andi Pratama
        follow_up_date:
          type: string
          format: date
          nullable: true
          example: '2026-03-05'
        notes:
          type: string
          nullable: true
          example: Jadwal diubah ke siang
        patient_allergies:
          type: string
          nullable: true
          example: 'Penicillin, seafood'
          description: Optional update for patient allergies from doctor's assessment
        patient_blood_type:
          type: string
          nullable: true
          enum:
            - A+
            - A-
            - B+
            - B-
            - AB+
            - AB-
            - O+
            - O-
          example: O+
          description: Optional update for patient blood type from doctor's assessment
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key